+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Implement Pushover emergency alert cancellation when check goes up
 - Add "The following checks are also down" section in Telegram notifications
 - Add "The following checks are also down" section in Signal notifications
+- Upgrade to django-compressor 3.0
 
 ### Bug Fixes
 - Fix report templates to not show the "started" status (show UP or DOWN instead)